fix: reduce value in filter complexity
